+++ lithops/serverless/backends/code_engine/code_engine.py
@@ -18,6 +18,7 @@
 import os
 import re
 import sys
+import base64
 import json
 import logging
 import urllib3
@@ -388,12 +389,55 @@ class CodeEngineBackend:
 
         return activation_id
 
+    def _create_container_registry_secret(self):
+        """
+        Create the container registry secret in the cluster
+        (only if credentials are present in config)
+        """
+        if not all(key in self.code_engine_config for key in ["docker_user", "docker_password"]):
+            return
+
+        logger.debug('Creating container registry secret')
+        docker_server = self.code_engine_config.get('docker_server', 'https://index.docker.io/v1/')
+        docker_user = self.code_engine_config.get('docker_user')
+        docker_password = self.code_engine_config.get('docker_password')
+
+        cred_payload = {
+            "auths": {
+                docker_server: {
+                    "Username": docker_user,
+                    "Password": docker_password
+                }
+            }
+        }
+
+        data = {
+            ".dockerconfigjson": base64.b64encode(
+                json.dumps(cred_payload).encode()
+            ).decode()
+        }
+
+        secret = client.V1Secret(
+            api_version="v1",
+            data=data,
+            kind="Secret",
+            metadata=dict(name="lithops-regcred", namespace=self.namespace),
+            type="kubernetes.io/dockerconfigjson",
+        )
+
+        try:
+            self.coreV1Api.create_namespaced_secret(self.namespace, secret)
+        except ApiException as e:
+            if e.status != 409:
+                raise e
+
     def _create_job_definition(self, image_name, runtime_memory, timeout):
         """
         Creates a Job definition
         """
-        jobdef_name = self._format_jobdef_name(image_name, runtime_memory)
+        self._create_container_registry_secret()
 
+        jobdef_name = self._format_jobdef_name(image_name, runtime_memory)
         jobdef_res = yaml.safe_load(ce_config.JOBDEF_DEFAULT)
         jobdef_res['metadata']['name'] = jobdef_name
         container = jobdef_res['spec']['template']['containers'][0]